+++ src/sybil/src/utils/web3/mod.rs
@@ -34,6 +34,7 @@ const TX_WAITING_TIMEOUT: u64 = 60 * 5;
 const TX_WAIT_DELAY: Duration = Duration::from_secs(3);
 
 pub mod evm_canister_transport_old;
+pub mod evm_canister_methods;
 pub mod promises;
 pub mod utils;
 
@@ -96,7 +97,7 @@ pub fn instance(
     chain_id: u64,
     rpcs_url: Option<Vec<String>>,
     _evm_rpc_canister: Principal,
-    max_response_bytes: Option<u64>,
+    _max_response_bytes: Option<u64>,
 
 ) -> Web3Instance<impl Transport> {
     // Switch between EVMCanisterTransport(calls go through emv_rpc canister) and ICHttp (calls go straight to the rpc)
@@ -114,7 +115,7 @@ pub fn batch_instance(
     chain_id: u64,
     rpcs_url: Option<Vec<String>>,
     evm_rpc_canister: Principal,
-    max_response_bytes: Option<u64>,
+    _max_response_bytes: Option<u64>,
 ) -> Web3Instance<Batch<impl BatchTransport>> {
     // Switch between EVMCanisterTransport(calls go through emv_rpc canister) and ICHttp (calls go straight to the rpc)
